runtime: reuse freed memory blocks on wasm
When compiling Go programs to WebAssembly, the memory allocation
strategy was neither releasing memory to the OS nor reusing blocks
freed by calls to runtime.sysFreeOS.

This CL unifies the plan9 and wasm memory management strategy since
both platforms use a linear memory space and do not have a mechanism
for returning memory blocks to the OS.

Fixes #59061
