+++ src/runtime/os_js.go
@@ -102,9 +102,11 @@ func mdestroy(mp *m) {
 }
 
 func osinit() {
+	// https://webassembly.github.io/spec/core/exec/runtime.html#memory-instances
+	physPageSize = 64 * 1024
+	initBloc()
 	ncpu = 1
 	getg().m.procid = 2
-	physPageSize = 64 * 1024
 }
 
 // wasm has no signals
